Chasm Proposal for Mammoth
This is my proposal for a new physical barrier for Mammoth. It is a large chasm that cuts through the center of the map. There are only two very narrow crossings with no railings, making it easy to contest a crossing from either side. The goal of this chasm is to focus players towards the mouth of the town. It is also to reduce the impetuous for early faction war by restricting the travel of large groups.

1. Why is this necessary?
People in the project have voiced concerns with the time to travel between the NCR base and the BoS base. These concerns I do not sure, but this is proposed to remedy the situation without creating a larger map size. This will cause points of contention and make it easier for delays in faction combat, and also focus crossings from one side of the map to the other.

2. What does this benefit the map itself?
No F13 map has had a large physical barrier like this before. Further, it showcases our multi-z and multi-z lighting, by allowing some natural light to filter down to the caves from above where the chasm is. People can fall on the paths into the caves below, creating much needed verticality in a very visible way.