The final wrap-up
I've got the map looking more or less how I wanted it to look, Spare for the hotel out in the front. It seems a bit dodgy in that location to me, but it can be moved later. The groundwork is there, so hopefully this UNOFFICIAL MAPPING FREEZE can thaw out, as hard as it is to thaw something out in a nuclear winter.