Going Nuclear: Powerplant and Village
Removed old bunker in a mountain. Replaced it with Powerplant (with Administration and Canteen). Improved Underground Village.
Made a Warehouse and Robotics Shop near Powerplant. Replaced and improved two buildings that were on their place before.
Reworked some "maintenance room" in a Sewers nearby Village that I made before. Fixed a stairs.
Fixed one unnecessary "roof helper" on a Police Station (spotted it in a last time when I opened the map).